\name{characiformes}
\docType{data}
\alias{characiformes}
\encoding{latin1}
\title{MORPHOMETRIC VARIABLES OF CHARACIFORMS}
\usage{data(characiformes)}
\description{Morphometric data of several species of 
Characiforms, as the length of the dorsal fin base (M12), body height (M11), etc. For details see Guisande et al. (2010).} 
\format{An array (matrix) with 31 columns: taxonomic data (order, family, genus and species) and 27 morphometric variables.}
\source{\url{http://www.ipez.es}.}
\references{
Guisande, C., Manjarrés-Hernández, A., Pelayo-Villamil, P., Granado-Lorencio, C., Riveiro, I., Acuña, A., Prieto-Piraquive, E., Janeiro, E., Matías, J.M., Patti, C., Patti, B., Mazzola, S., Jiménez, S., Duque, V. & Salmerón, F. (2010) IPez: An expert system for the taxonomic identification of fishes based on machine learning techniques. \emph{Fisheries Research}, 102, 240-247.}
\keyword{characiformes}
